+++ Jellyfin.Plugin.LibraryManager/Model/ErrorResponse.cs
@@ -1,0 +1,20 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Text.Json.Serialization;
+using System.Threading.Tasks;
+
+namespace Jellyfin.Plugin.LibraryManager.Model
+{
+    /// <summary>
+    /// The error response.
+    /// </summary>
+    public class ErrorResponse
+    {
+        /// <summary>
+        /// Gets or sets the error message.
+        /// </summary>
+        public string Message { get; set; }
+    }
+}